refactor: split control panel into focused sub-components
Extract 7 presentational sub-components from ControlPanel.tsx (981→533 lines):
- ModeSelector: Serial/TCP toggle
- SerialConfigPanel: port, baud, databits, parity, stopbits, framing
- NetworkConfigPanel: host + port inputs
- PresetToolbar: preset management (update/revert/copy/detach/save-as)
- ProjectActions: import/export/AI generator buttons
- ConnectionButton: connect/disconnect button
- FramingModal: framing strategy portal modal

Container keeps all state, effects, and handlers. Sub-components are
pure presentational receiving state/callbacks via props.

via [Happy](https://happy.engineering)
